Tier 2.2: Extract SpriteUtils and reduce GameSprite friend declarations
Extract static Decompress() and ColorizeTemplatePixels() methods to
SpriteUtils namespace in sprite_utils.h/cpp. GameSprite delegates
to SpriteUtils for backward compatibility. Remove 4 unnecessary
friend declarations (GraphicManager, TextureGarbageCollector,
TooltipDrawer, SpritePreloader) that only accessed public members.
Only GraphicsAssembler and SpriteIconGenerator remain as friends.
